requirements.txt
streamlit
streamlit-option-menu
streamlit-extras
oracledb
pandas
numpy
plotly
st-pages
requests
psycopg2-binary
